+++ src/_2024/Day06/Part01.cs
@@ -25,7 +25,7 @@ public class Part01 : PuzzleSolver<long>
             if (!nextPos.InBounds(0, 0, map[0].Length - 1, map.Length - 1))
                 break;
             
-            if (map[nextPos.Y][nextPos.X] == '#')
+            if (map[(int)nextPos.Y][(int)nextPos.X] == '#')
                 queue.Enqueue((curPos, Point.RotateRight90(dir)));
             else
             {
